Fix: Quiz 엔티티의 answerRate 필드를 Double 타입으로 수정

+++ src/main/kotlin/com/youquiz/quiz/domain/Quiz.kt
@@ -14,7 +14,7 @@ class Quiz(
     val solution: String,
     val writer: User,
     val chapterId: String,
-    var answerRate: Long,
+    var answerRate: Double,
     var correctCount: Long,
     var incorrectCount: Long,
 ) {
